+++ src/demo/ts/ephox/alloy/demo/DialogDemo.ts
@@ -54,13 +54,19 @@ export default (): void => {
           innerHtml: 'Wait for 5 seconds'
         },
         action: (comp) => {
-          ModalDialog.setBusy(dialog, (dlg, boundsStyles, busyBehaviours) => ({
+          ModalDialog.setBusy(dialog, (dlg, busyBehaviours) => ({
             dom: {
               tag: 'div',
               attributes: {
                 'aria-label': 'Sharks and Dolphins'
               },
-              styles: boundsStyles
+              styles: {
+                position: 'absolute',
+                left: '0px',
+                right: '0px',
+                bottom: '0px',
+                top: '0px'
+              }
             },
             components: [
               {
@@ -97,7 +103,7 @@ export default (): void => {
 
   const dialog = GuiFactory.build(
     ModalDialog.sketch({
-      dom: DomFactory.fromHtml('<div class="mce-container mce-panel mce-floatpanel mce-window mce-in"></div>'),
+      dom: DomFactory.fromHtml('<div class="mce-container mce-panel mce-floatpanel mce-window mce-in" style="position: relative;"></div>'),
       components: [
         Container.sketch({
           dom: DomFactory.fromHtml('<div class="mce-reset" role="application"></div>'),