+++ src/main/ts/ephox/alloy/api/events/AlloyEvents.ts
@@ -153,6 +153,10 @@ const stopper = function <T extends EventFormat>(name): AlloyEventKeyAndHandler<
   });
 };
 
+const runOnSource = function <T extends EventFormat>(name: string, f: EventRunHandler<T>): AlloyEventKeyAndHandler<T> {
+  return runOnSourceName(name)(f);
+}
+
 const runOnAttached = runOnSourceName(SystemEvents.attachedToDom());
 const runOnDetached = runOnSourceName(SystemEvents.detachedFromDom());
 const runOnInit = runOnSourceName(SystemEvents.systemInit());
@@ -164,6 +168,7 @@ export {
   runActionExtra,
   runOnAttached,
   runOnDetached,
+  runOnSource,
   runOnInit,
   runOnExecute,